requirements.txt
# Core dependencies
cobra>=0.26.0
pandas>=1.3.0
numpy>=1.20.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=5.4.0

# Progress bars
tqdm>=4.60.0

# Optional: Advanced visualizations
# matplotlib-venn>=0.11.0
# upsetplot>=0.6.0
# networkx>=2.6.0

# Optional: Solvers (install separately)
# cplex
# gurobipy

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=6.0.0
# pytest-cov>=2.12.0
# black>=21.0.0
# flake8>=3.9.0
# mypy>=0.900
